pyproject.toml
[project]
name = "home-device-bridge"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0,<3.0.0",
    "httpx>=0.28.0",
    "starlette>=0.36.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
